+++ Fumen Utils_files/fumenutil/modified-combineFumen.js
@@ -1,5 +1,5 @@
 const { decoder, encoder } = require('tetris-fumen');
-import { LineTerminator } from "./global-utils.js";
+import { LineTerminator } from "../global-utils.js";
 
 export default function combineFumen() {
     var input = document.getElementById('input').value.replace(/[Ddm]115@/gm,'v115@')
