+++ Fumen Utils_files/fumenutil/modified-mirrorFumen.js
@@ -1,5 +1,5 @@
 const { decoder, encoder } = require('tetris-fumen');
-import { getDelimiter, LineTerminator } from './global-utils.js'
+import { getDelimiter, LineTerminator } from '../global-utils.js'
 
 // var colorMapping = { //unused
 //     "S": 7,
