+++ Fumen Utils_files/lib/GIFEncoder.js
