+++ Fumen Utils_files/lib/NeuQuant.js
