+++ index.html
@@ -12,27 +12,15 @@
 
 <!--EXTERNAL SCRIPTS-->
 <script type = "module" src = "./base.js"></script>
-<script src = "./Fumen Utils_files/tetris-fumen.js"></script>
-<!-- <script type = "module" src = "./Fumen Utils_files/modified-glueFumen.js"></script> -->
-<!-- <script type = "module" src = "./Fumen Utils_files/modified-unglueFumen.js"></script>
-<script type = "module" src = "./Fumen Utils_files/modified-mirrorFumen.js"></script>
-<script type = "module" src = "./Fumen Utils_files/modified-combineFumen.js"></script>
-<script type = "module" src = "./Fumen Utils_files/modified-splitFumen.js"></script>
-<script type = "module" src = "./Fumen Utils_files/modified-removeComments.js"></script> -->
-<!-- <script type = "module" src = "./Fumen Utils_files/modified-fumen-canvas.js"></script>
-<script type = "module" src = "./Fumen Utils_files/modified-fumen-render.js"></script> -->
-<!-- <script src = "./Fumen Utils_files/board-render.js"></script> -->
-<script type = "module" src = "./Fumen Utils_files/event-listeners.js"></script>
-<script src = "./Fumen Utils_files/jszip.min.js"></script>
-<script src = "./Fumen Utils_files/jszip-utils.min.js"></script>
-<script src = "./Fumen Utils_files/FileSaver.min.js"></script>
-<script src = "./Fumen Utils_files/LZWEncoder.js"></script>
-<script src = "./Fumen Utils_files/NeuQuant.js"></script>
-<script src = "./Fumen Utils_files/GIFEncoder.js"></script>
-<!-- <script type = "module" src = "./Fumen Utils_files/gifenc/gifenc.esm.js"></script> -->
-<script type = "module" src = "./Fumen Utils_files/b64.js"></script>
-<script src = "./Fumen Utils_files/mousetrap.min.js"></script>
-<script src = "./Fumen Utils_files/mousetrap-bind-dictionary.min.js"></script>
+<script src = "./Fumen Utils_files/lib/tetris-fumen.js"></script>
+<script src = "./Fumen Utils_files/lib/jszip.min.js"></script>
+<script src = "./Fumen Utils_files/lib/jszip-utils.min.js"></script>
+<script src = "./Fumen Utils_files/lib/FileSaver.min.js"></script>
+<script src = "./Fumen Utils_files/lib/LZWEncoder.js"></script>
+<script src = "./Fumen Utils_files/lib/NeuQuant.js"></script>
+<script src = "./Fumen Utils_files/lib/GIFEncoder.js"></script>
+<script src = "./Fumen Utils_files/lib/mousetrap.min.js"></script>
+<script src = "./Fumen Utils_files/lib/mousetrap-bind-dictionary.min.js"></script>
 <script src = "https://cdn.jsdelivr.net/npm/html2canvas@1.0.0-rc.5/dist/html2canvas.min.js"></script>
 
 	<body id = "main">
